cleaning up miner class, moving blacklist/priority to base
