+++ neurons/miner.py
@@ -47,13 +47,28 @@ class Miner(BaseMinerNeuron):
 
     def __init__(self, config=None):
         super(Miner, self).__init__(config=config)
+        bt.logging.info("Attaching forward function to miner axon.")
+        self.axon.attach(
+            forward_fn=self.forward_image,
+            blacklist_fn=self.blacklist_image,
+            priority_fn=self.priority_image,
+        ).attach(
+            forward_fn=self.forward_video,
+            blacklist_fn=self.blacklist_video,
+            priority_fn=self.priority_video,
+        )
+        bt.logging.info(f"Axon created: {self.axon}")
+
+        bt.logging.info("Loading image detection model if configured")
         self.load_image_detector()
+        bt.logging.info("Loading video detection model if configured")
         self.load_video_detector()
 
     def load_image_detector(self):
         if str(self.config.neuron.image_detector).lower() == 'none' or \
             str(self.config.neuron.image_detector_config).lower() == 'none':
             bt.logging.warning("No image detector configuration provided, skipping.")
+            self.image_detector = None
             return
 
         if self.config.neuron.image_detector_device == 'auto':
@@ -71,6 +86,7 @@ class Miner(BaseMinerNeuron):
         if str(self.config.neuron.video_detector).lower() == 'none' or \
             str(self.config.neuron.video_detector_config).lower() == 'none':
             bt.logging.warning("No video detector configuration provided, skipping.")
+            self.video_detector = None
             return
 
         if self.config.neuron.video_detector_device == 'auto':
@@ -98,19 +114,19 @@ class Miner(BaseMinerNeuron):
             bt.Synapse: The synapse object with the 'predictions' field populated with a list of probabilities
 
         """
-        bt.logging.info("Received image challenge!")
-        try:
-            image_bytes = base64.b64decode(synapse.image)
-            image = Image.open(io.BytesIO(image_bytes))
-
-            pred = self.deepfake_detector(image)
-            synapse.prediction = pred
-
-        except Exception as e:
-            bt.logging.error("Error performing inference")
-            bt.logging.error(e)
-
-        bt.logging.info(f"PREDICTION: {synapse.prediction}")
+        if self.image_detector is None:
+            bt.logging.info("Image detection model not configured; skipping image challenge")
+        else:
+            bt.logging.info("Received image challenge!")
+            try:
+                image_bytes = base64.b64decode(synapse.image)
+                image = Image.open(io.BytesIO(image_bytes))
+                pred = self.image_detector(image)
+                synapse.prediction = pred
+            except Exception as e:
+                bt.logging.error("Error performing inference")
+                bt.logging.error(e)
+            bt.logging.info(f"PREDICTION: {synapse.prediction}")
         return synapse
 
     async def forward_video(
@@ -126,122 +142,25 @@ class Miner(BaseMinerNeuron):
             bt.Synapse: The synapse object with the 'predictions' field populated with a list of probabilities
 
         """
-        bt.logging.info("Received video challenge!")
-        try:
-            frames = decode_video_synapse(synapse)
-            pred = self.deepfake_detector({'image': frames})
-            synapse.prediction = torch.max(pred)
-
-        except Exception as e:
-            bt.logging.error("Error performing inference")
-            bt.logging.error(e)
-
-        bt.logging.info(f"PREDICTION: {synapse.prediction}")
+        if self.video_detector is None:
+            bt.logging.info("Video detection model not configured; skipping video challenge")
+        else:
+            bt.logging.info("Received video challenge!")
+            try:
+                frames = decode_video_synapse(synapse)
+                pred = self.video_detector({'image': frames})
+                synapse.prediction = torch.max(pred)
+            except Exception as e:
+                bt.logging.error("Error performing inference")
+                bt.logging.error(e)
+            bt.logging.info(f"PREDICTION: {synapse.prediction}")
         return synapse
-
-    async def blacklist(
-        self, synapse: bt.Synapse
-    ) -> typing.Tuple[bool, str]:
-        """
-        Determines whether an incoming request should be blacklisted and thus ignored. Your implementation should
-        define the logic for blacklisting requests based on your needs and desired security parameters.
-
-        Blacklist runs before the synapse data has been deserialized (i.e. before synapse.data is available).
-        The synapse is instead contructed via the headers of the request. It is important to blacklist
-        requests before they are deserialized to avoid wasting resources on requests that will be ignored.
-
-        Args:
-            synapse (bt.Synapse): A synapse object constructed from the headers of the incoming request.
-
-        Returns:
-            Tuple[bool, str]: A tuple containing a boolean indicating whether the synapse's hotkey is blacklisted,
-                            and a string providing the reason for the decision.
-
-        This function is a security measure to prevent resource wastage on undesired requests. It should be enhanced
-        to include checks against the metagraph for entity registration, validator status, and sufficient stake
-        before deserialization of synapse data to minimize processing overhead.
-
-        Example blacklist logic:
-        - Reject if the hotkey is not a registered entity within the metagraph.
-        - Consider blacklisting entities that are not validators or have insufficient stake.
-
-        In practice it would be wise to blacklist requests from entities that are not validators, or do not have
-        enough stake. This can be checked via metagraph.S and metagraph.validator_permit. You can always attain
-        the uid of the sender via a metagraph.hotkeys.index( synapse.dendrite.hotkey ) call.
-
-        Otherwise, allow the request to be processed further.
-        """
-        if synapse.dendrite is None or synapse.dendrite.hotkey is None:
-            bt.logging.warning("Received a request without a dendrite or hotkey.")
-            return True, "Missing dendrite or hotkey"
-
-        # TODO(developer): Define how miners should blacklist requests.
-        uid = self.metagraph.hotkeys.index(synapse.dendrite.hotkey)
-        if (
-            not self.config.blacklist.allow_non_registered
-            and synapse.dendrite.hotkey not in self.metagraph.hotkeys
-        ):
-            # Ignore requests from un-registered entities.
-            bt.logging.trace(
-                f"Blacklisting un-registered hotkey {synapse.dendrite.hotkey}"
-            )
-            return True, "Unrecognized hotkey"
-
-        if self.config.blacklist.force_validator_permit:
-            # If the config is set to force validator permit, then we should only allow requests from validators.
-            if not self.metagraph.validator_permit[uid]:
-                bt.logging.warning(
-                    f"Blacklisting a request from non-validator hotkey {synapse.dendrite.hotkey}"
-                )
-                return True, "Non-validator hotkey"
-
-        bt.logging.trace(
-            f"Not Blacklisting recognized hotkey {synapse.dendrite.hotkey}"
-        )
-        return False, "Hotkey recognized!"
 
     async def blacklist_image(self, synapse: ImageSynapse) -> typing.Tuple[bool, str]:
         return await self.blacklist(synapse)
 
     async def blacklist_video(self, synapse: VideoSynapse) -> typing.Tuple[bool, str]:
         return await self.blacklist(synapse)
-
-    async def priority(self, synapse: bt.Synapse) -> float:
-        """
-        The priority function determines the order in which requests are handled. More valuable or higher-priority
-        requests are processed before others. You should design your own priority mechanism with care.
-
-        This implementation assigns priority to incoming requests based on the calling entity's stake in the metagraph.
-
-        Args:
-            synapse (bt.Synapse): The synapse object that contains metadata about the incoming request.
-
-        Returns:
-            float: A priority score derived from the stake of the calling entity.
-
-        Miners may recieve messages from multiple entities at once. This function determines which request should be
-        processed first. Higher values indicate that the request should be processed first. Lower values indicate
-        that the request should be processed later.
-
-        Example priority logic:
-        - A higher stake results in a higher priority value.
-        """
-        if synapse.dendrite is None or synapse.dendrite.hotkey is None:
-            bt.logging.warning("Received a request without a dendrite or hotkey.")
-            return 0.0
-
-        # TODO(developer): Define how miners should prioritize requests.
-        caller_uid = self.metagraph.hotkeys.index(
-            synapse.dendrite.hotkey
-        )  # Get the caller index.
-
-        prirority = float(
-            self.metagraph.S[caller_uid]
-        )  # Return the stake as the priority.
-        bt.logging.trace(
-            f"Prioritizing {synapse.dendrite.hotkey} with value: ", prirority
-        )
-        return prirority
 
     async def priority_image(self, synapse: ImageSynapse) -> float:
         return await self.priority(synapse)
