+++ neurons/validator.py
@@ -158,4 +158,4 @@ if __name__ == "__main__":
     with Validator() as validator:
         while True:
             bt.logging.info(f"Validator running | uid {validator.uid} | {time.time()}")
-            time.sleep(30)
+            time.sleep(5)
